restructured device functionality regarding messages
