+++ src/main/java/zork/enviromnment/Device.java
@@ -6,6 +6,7 @@ import akka.actor.AbstractActor;
 import akka.actor.Props;
 import akka.event.Logging;
 import akka.event.LoggingAdapter;
+import zork.enviromnment.messages.DeviceFunctionality;
 import zork.enviromnment.messages.DeviceLifecycle.DeviceRegistered;
 import zork.enviromnment.messages.DeviceLifecycle.RequestTrackDevice;
 
@@ -22,42 +23,6 @@ public class Device extends AbstractActor {
 
   public static Props props(String groupId, String deviceId) {
     return Props.create(Device.class, () -> new Device(groupId, deviceId));
-  }
-
-  public static final class RecordTemperature {
-    final long requestId;
-    final double value;
-
-    public RecordTemperature(long requestId, double value) {
-      this.requestId = requestId;
-      this.value = value;
-    }
-  }
-
-  public static final class TemperatureRecorded {
-    public final long requestId;
-
-    public TemperatureRecorded(long requestId) {
-      this.requestId = requestId;
-    }
-  }
-
-  public static final class ReadTemperature {
-    public final long requestId;
-
-    public ReadTemperature(long requestId) {
-      this.requestId = requestId;
-    }
-  }
-
-  public static final class RespondTemperature {
-    public final long requestId;
-    public final Optional<Double> value;
-
-    public RespondTemperature(long requestId, Optional<Double> value) {
-      this.requestId = requestId;
-      this.value = value;
-    }
   }
 
   Optional<Double> lastTemperatureReading = Optional.empty();
@@ -88,12 +53,12 @@ public class Device extends AbstractActor {
         );
       }
     })
-    .match(RecordTemperature.class, r -> {
+    .match(DeviceFunctionality.RecordTemperature.class, r -> {
       log.info("Recorded temperature reading {} with {}", r.value, r.requestId);
       lastTemperatureReading = Optional.of(r.value);
-      getSender().tell(new TemperatureRecorded(r.requestId), getSelf());
-    }).match(ReadTemperature.class, r -> {
-      getSender().tell(new RespondTemperature(r.requestId, lastTemperatureReading), getSelf());
+      getSender().tell(new DeviceFunctionality.TemperatureRecorded(r.requestId), getSelf());
+    }).match(DeviceFunctionality.ReadTemperature.class, r -> {
+      getSender().tell(new DeviceFunctionality.RespondTemperature(r.requestId, lastTemperatureReading), getSelf());
     }).build();
   }
 }
